% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Cluster record IDs into terminal nodes}
\usage{
cluster(x, step = 0, ...)
}
\arguments{
\item{x}{}
}
\value{
A matrix containing the record number and the terminal node
         it corresponds to
}
\description{
Produce clusters based on terminal nodes
of tree
}
