% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbrpart.control.R
\name{dbrpart.control}
\alias{dbrpart.control}
\title{Control for dbrpart fits}
\usage{
dbrpart.control(minSplit = 20, minBucket = round(minSplit/3), c("default","mean","median","mode"), 
         maxDepth = 20, distanceAdjust=0, trace = 3, 
         stopMethod = c("default","cv","permutation","none"), nSims=5000, signifLevel=0.05,
         pCenters=1, combinationPower = 2)
}
\arguments{
\item{minSplit}{the minimum number of observations that must exist in a node in order for a split to be attempted.}

\item{minBucket}{the minimum number of observations in any terminal node}

\item{classMethod}{the method for classifying members in a node}

\item{maxDepth}{the maximum depth of the tree}

\item{distanceAdjust}{a penalty when deciding bubble center based on total distance to other observations.}

\item{trace}{the level of information to print while dbrpart runs}

\item{stopMethod}{the method for determining when stopping occurs}

\item{nSims}{Number of permutations to perform in permutation stopping criterion. 0 for cross-validation pruning, 1 for permutation tests, 2 for no stopping}

\item{signifLevel}{Significance level for permutation stopping criterion}

\item{pCenters}{}

\item{combinationPower}{}
}
\description{
Control for dbrpart fits
}
