% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbrpart.R
\name{dbrpart}
\alias{dbrpart}
\title{Recursive composite k-nearest neighbour partitioning}
\usage{
dbrpart(
  target,
  factorList = NULL,
  numericList = NULL,
  distanceList = NULL,
  distanceCombinations = NULL,
  priors = NULL,
  alteredPriors = FALSE,
  loss = NULL,
  controls = NULL,
  soma.controls = NULL
)
}
\arguments{
\item{target}{Response Variable. Must be either a numeric, logical or factor vector, or a distance matrix.}

\item{factorList}{}

\item{numericList}{}

\item{distanceList}{A list of distance matrices. rcknnpart will partition using these features.}

\item{distanceCombinations}{}

\item{priors}{Prior distribution for categorical outcome frequency. Ignored if target is not a factor.}

\item{alteredPriors}{}

\item{loss}{Loss matrix for categorical outcome frequency. Ignored if target is not a factor.}

\item{controls}{}

\item{soma.controls}{}
}
\value{
A list containing the following

\itemize{

If cross-validation is being used:

\item `a` - the alpha complexity thresholds for each prune
\item `b` - beta, the geometric mean between successive pruning values
\item `nTerm` - the number of terminal nodes at pruning step
\item `model_risk` - the average error at each pruning step
\item `risk` - the sum of errors at each pruning step
\item `cv_risk` - the sum of errors at each pruning step under cross-validation
\item `classTree` - the terminal node that each observation belongs to at each pruning step
\item `prediction` - the predicted value of each observation at each pruning step (under the model)
\item `details`
\item `cv_prediction` - the predicted value of each observation at each pruning step (under cross-validation)
\item `cv_details`

For all methods, the following:

\item `tree` -  a list containing the details of each node in the tree:
\itemize{
    \item `label` - the label for the node. This is currently zero indexed and should be changed.
    \item `parent` - the label for the node's parent. Currently, this doesn't contain anything and should probably be removed.
    \item `probability` - the probability of a random patient being in this node
    \item `groupMembers` - a list of observations in this node. This is currently zero indexed and should be changed.
    \item  `groupSize` - the count of observations in each node.
    \item `risk` - the ``risk'' of the node. This changes depending on the target:
    \itemize{
        \item If the target is nominal or ordinal, then it contains the 
        \item If the target is scalar, then it should contain 
        \item If the target is distance-based,, then
    }
    \item `pruneAtStep`
    \item `impurity`
    \item `classif`
    \item `mean`
    \item `terminal`
    \item `alpha`
    \item `splitMethod`
    \item `bestSplitIndex`
    \item `splitName`
    \item `splitUtility`
    \item `goodnessOfFit`
    \item `center`
    \item `radius`
    \item `pVal`
    \item `inChild`
    \item `outChild`
}



\item `y` - the target variable.
\item `featureNames`
\item `classNames`
\item `mode`
\item `controls`
\item `foldList`
\item `bestPruneStep`
}
}
\description{
Variation on CART using distance-based paritioning
}
