% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCpTable.R
\name{getCpTable}
\alias{getCpTable}
\title{Get Complexity Parameter Table for a dbrpart fit}
\usage{
getCpTable(x)
}
\arguments{
\item{x}{an object of class \code{"Dbrpart"}}
}
\value{
a list containing
\itemize{
    \item \code{cpTable}
    \item \code{index} - the row of cpTable corresponding to the best fit using the 1-SE rule 
}
}
\description{
Get Complexity Parameter Table for a dbrpart fit
}
